discoverd: Replace with discoverd2
