+++ discoverd/daemon.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/flynn/flynn/Godeps/_workspace/src/github.com/coreos/go-etcd/etcd"
-	"github.com/flynn/flynn/discoverd2/server"
+	"github.com/flynn/flynn/discoverd/server"
 	"github.com/flynn/flynn/pkg/attempt"
 )
