+++ discoverd/server/dns_test.go
@@ -10,7 +10,7 @@ import (
 
 	. "github.com/flynn/flynn/Godeps/_workspace/src/github.com/flynn/go-check"
 	"github.com/flynn/flynn/Godeps/_workspace/src/github.com/miekg/dns"
-	"github.com/flynn/flynn/discoverd2/client"
+	"github.com/flynn/flynn/discoverd/client"
 )
 
 type DNSSuite struct {
