+++ discoverd/server/http_test.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/flynn/flynn/Godeps/_workspace/src/github.com/coreos/go-etcd/etcd"
 	. "github.com/flynn/flynn/Godeps/_workspace/src/github.com/flynn/go-check"
+	"github.com/flynn/flynn/discoverd/client"
 	"github.com/flynn/flynn/discoverd/testutil/etcdrunner"
-	"github.com/flynn/flynn/discoverd2/client"
 )
 
 var _ = Suite(&HTTPSuite{})
